+++ banto/src/main/java/com/example/banto/Users/UserRepository.java
@@ -1,13 +1,14 @@
 package com.example.banto.Users;
 
+import java.awt.print.Pageable;
+import java.util.List;
 import java.util.Optional;
 
+import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.banto.Users.Users;
-
 @Repository
-public interface UserRepository extends JpaRepository<Users, Integer> {
+public interface UserRepository extends JpaRepository<Users, Long> {
     Optional<Users> findByEmail(String email);
 }
